Change device detection to LIS
UART is always busy by the driver, so we can't check it using existing
methods

LIS3MDL however is normal, and just in a slightly different location. It
might not always stay there, but I think we need to just start adding
versions of NaviCube to handle this in the future.